+++ python/bartons/indicators/lib/ker.py
@@ -2,8 +2,8 @@ import polars as pl
 
 from polars.plugins import register_plugin_function
 
-from ..prelude import PLUGIN_PATH, wrap_src_indicator
-from ..typing import IntoExprColumn
+from ...prelude import PLUGIN_PATH, wrap_src_indicator
+from ...typing import IntoExprColumn
 
 __all__ = ("KER",)