+++ python/bartons/indicators/lib/macd.py
@@ -1,7 +1,7 @@
 import polars as pl
 
-from ..prelude import wrap_src_indicator
-from ..typing import IntoExprColumn
+from ...prelude import wrap_src_indicator
+from ...typing import IntoExprColumn
 from .ema import EMA
 
 __all__ = ("MACD",)