+++ python/bartons/indicators/lib/sar.py
@@ -1,8 +1,8 @@
 import polars as pl
 from polars.plugins import register_plugin_function
 
-from ..prelude import PLUGIN_PATH, wrap_indicator
-from ..typing import IntoExprColumn
+from ...prelude import PLUGIN_PATH, wrap_indicator
+from ...typing import IntoExprColumn
 
 __all__ = ("SAR",)